msm: kgsl: Overhaul coresight
Do a major reorganization of the coresight code. Move all the
coresight definitions to their own files that are conditional on
CONFIG_QCOM_KGSL_CORESIGHT. Next, add the regmap for "cx_dbgc" so that
we can get rid of the single purpose cx_dgbc registers and replace them
with the standard register read/write functions. Then finish making the
coresight functions more generic and have the individual targets
register themselves with coresight instead of keeping a pointer in the
gpudev struct.

Change-Id: Ic479934ef2ec6413cbac7498206d0202e7f76d5a
Signed-off-by: Nitheesh Muthuraj <quic_nmuthuraj@quicinc.com>
Signed-off-by: Mohammed Mirza Mandayappurath Manzoor <quic_mmandaya@quicinc.com>