msm: kgsl: Add notifier for CX gdsc collapse
During slumber sequence, kgsl polls (sleeps in between) for CX gdsc
to go down. This has few problems:
1. Driver does significant amount of polling which could waste CPU
   cycles.
2. It is possible that CX can go down while kgsl is sleeping and
   could be back to on state when it checks again.
3. Poll happens inside device mutex which could block other things.

In this change, kgsl registers a notifier for cx regulator events.
Instead of polling for cx to go down, we will wait for the cx collapse
notifier event. On receiving the notification, kgsl polls cx gdsc
register to ensure that CX is collapsed. With this change, device
mutex is not taken while polling allowing other operations to go
through.

Change-Id: I760a900ed94ed98f63b83746dbafabd6d4d7a615
Signed-off-by: Kamal Agrawal <quic_kamaagra@quicinc.com>