msm: kgsl: Use vlvl instead of hlvl in DT
Currently, we use hlvl value in device tree to identify the cx minimum
vote for each opp corner. But, hlvl values are not guaranteed to
represent a particular voltage corner throughout the life time of the
SoC. Instead of this, we are supposed to use vlvl value in device tree
and then identify the corresponding hlvl value with the help of
command DB during runtime.

Also, rename dt property from "qcom,min-cx-level" to "qcom,cx-level" to
be more concise.

Change-Id: I0c811114614422b99ebb9daeee355a53292a7412
Signed-off-by: Akhil P Oommen <quic_akhilpo@quicinc.com>