msm: kgsl: Do not stall on GPU pagefault if hwsched is enabled
If there is a GPU pagefault and stall on fault is enabled for the
GPU context bank, the stall propagates to the GMU context bank as
well because part of the path to memory is shared between GPU and
the GMU. As a result, the GMU can hang on any arbitrary memory
transaction resulting in failed snapshot and recovery.
To work around this issue, configure the GPU to terminate the
faulting transaction and any subsequent transactions instead of
stalling when there is a GPU pagefault.

Change-Id: Icca207b43e06887f477091ef985ed40e0e62d8c8
Signed-off-by: Sushmita Susheelendra <quic_ssusheel@quicinc.com>