msm: kgsl: Do not process HFI queues in F2H daemon during init sequence
If the F2H daemon runs and finds messages in the HFI queues,
it will read them off the queue and process them. This should
not happen during the init sequence while interrupts on the
HFI message queue are not yet enabled. Otherwise, the F2H
daemon will read messages off the queue, leaving the polling
thread confused.

Change-Id: I8331e535fac37d3ac4c9884529ed603f4882730c
Signed-off-by: Sushmita Susheelendra <ssusheel@codeaurora.org>
Signed-off-by: Puranam V G Tejaswi <quic_pvgtejas@quicinc.com>