msm: kgsl: Remove attribute groups associated with gfx funnel
Coresight driver provides a sysfs node for funnel devices to show the
status of the funnel. Since this involves coresight driver reading the
control register of the funnel in cx domain, a NOC error would occur
when the device is in slumber. As this node is not required for gfx
funnel, remove the sysfs attribute groups associated with the funnel.

Change-Id: I3b8048dc1242f744df4c51696631f1e188e14c68
Signed-off-by: Mohammed Mirza Mandayappurath Manzoor <quic_mmandaya@quicinc.com>