msm: kgsl: Allow single fw for secure ucode in Gen7_2_0
So far, we have used the split version of secure ucode (.mdt) which
requires loading of multiple files from userpace. Obviously, it is slow
and has higher chance of failure due to multiple access to userspace.
We can move to the single fw version of secure ucode (.mbn) for newer
gpus to avoid these issues.

On an average, load time with split firmware is around 170ms but with
single firmware binary, it reduces to around 60ms.

Change-Id: Icb781f74192f3df4a92c99ef2da2f96ddcddd25a
Signed-off-by: Akhil P Oommen <quic_akhilpo@quicinc.com>
Signed-off-by: Kamal Agrawal <quic_kamaagra@quicinc.com>