msm: kgsl: Ignore pending_free flag for process restore
We check pending_free flag when pages are reclaimed and
when they restored. This is problematic because it is
possible to set pending_free flag after reclaim but before
restore. This will result in the reclaimed page not being
restored, thus causing page faults. Ignore pending_free
flag for process restore so that every page that is reclaimed
would be pinned back regardless of pending_free flag.

Change-Id: I6adf12021f8c8b8225ba55c85406fff1c11d927a
Signed-off-by: Rohan Sethi <quic_rohsethi@quicinc.com>