msm: kgsl: Program correct control bits for gen7_2_x PDC sequence
Control bits in the PDC sequence are used to control the HW branching
logic. Since the PDC sequence is shared, we need to program the register
with the start address to the branches that are applicable for the GPU.
The GPU needs to set events for PDC_BRANCH_NO_WAKEUP_TIMER and
PDC_BRANCH_SLEEP_TCS0 for gen7_2_x.

Change-Id: I87ec667ba251dfbe6930aedc212871eaa1c308c9
Signed-off-by: Mohammed Mirza Mandayappurath Manzoor <quic_mmandaya@quicinc.com>