msm: kgsl: Handle F2H_MSG_MEM_ALLOC alignment request
Take in to account the alignment parameter when computing the
GMU VA for a particular mem alloc packet. Since we are adding new
member at the very end of the hfi_mem_alloc_reply_cmd structure,
it is safe to use the new structure in the reply, even if
the GMU FW expects the old structure.

Change-Id: I552be29567fea7b6b2fa9285fbc1f2963850622d
Signed-off-by: Harshdeep Dhatt <quic_hdhatt@quicinc.com>