msm: kgsl: Fix HFI teardown sequence
The HFI irq handler only clears interrupts that it expects
will fire in order to not race with the GMU init sequence.
When HFI is stopped, HFI message queue interrupts are no
longer expected to fire. In an error scenario, if the GMU
triggers an interrupt to the message queue after this point,
the irq handler will not clear it leading to an irq storm.
Therefore disable GMU interrupts before we stop HFI.
Process the GMU message queue one last time before resetting
the HFI queues to not lose any messages that the GMU sent
since the interrupts were disabled. After this point in time,
HFI queues will not be read before GMU reboot.

Change-Id: I5d7b278ed246d7d139d4c3e5a099a0e803846d17
Signed-off-by: Sushmita Susheelendra <quic_ssusheel@quicinc.com>