msm: kgsl: Update logic to determine ib address overlap
Snapshot structure's ib1size represents the remaining number of
dwords to be processed by GPU. IB address overlap is independent
of the remaining dwords to be processed. Thus, update the logic
to determine if active IB address read from register falls within
IB buffer range submitted to RB.

Currently, if the remaining number of dwords to be processed is 0,
the validation condition fails and active IB is assumed to be not
present in RB. This change helps to fix the above scenario.

Change-Id: I26a36f450221e74ac61f10b8f88791aaf984e934
Signed-off-by: Kamal Agrawal <quic_kamaagra@quicinc.com>