ipa: Fix ioctl buffer allocation due to increased header size
The max header size changed from 64 to 255, which caused increase
in size of structs containing headers (struct ipa_hdr_add). Fixing the stack buffer
allocation size in the ioctl to avoid buffer overflow.

Note for the future: This function allocates number of buffers
on the stack, but uses only one on each call. It would be better
to refactor it, to use only one buffer, in order to minimize
stack allocations.

Change-Id: Id79efe7901ebf85dab4c0ba09cc7bb852cf22a97
Signed-off-by: Ilia Lin <quic_ilial@quicinc.com>
Signed-off-by: Chaitanya Pratapa <quic_cpratapa@quicinc.com>