msm: camera: cpas: Resolve integer endianness
(uint32_t *) &cpas_core->ahb_bus_client.common_data.bw_pair[i].ab rely
on the endianness of the device. By down casting a uint64_t* to a
uint32_t* we're writing to the first 32 bits of the int which are the
most significant 4 bytes or the least significant four bytes depending
on endianness.
To resolve it, take a temporary uint32_t variable to pass to the
of_property_read_u32_index function.

CRs-Fixed: 3314763
Change-Id: Ic6e2668e9d9f9e80b97a3a82467d02597bb51c44
Signed-off-by: Dharmender Sharma <quic_dharshar@quicinc.com>