msm: kgsl: Disable CP i-cache prefetch for gen7_2_x
It is observed in rare cases that CP takes longer to boot during IFPC
exit when CP Icache prefetch is enabled, resulting in gmu timeout.
So, CP Icache prefetch is disabled on few gen7 variants.
Add gen7_2_x variants also to this list as it helps in reducing IFPC
exit latency.

Change-Id: Iac2427f5349e19987e2e41d2cfd11a3b7c0e2320
Signed-off-by: Raviteja Narayanam <quic_ravnar@quicinc.com>