msm: kgsl: Fix time of check and time of use at few places
Firmware updates the read index value. The value of read index
may change between the time of validation and time of use, if
it is directly read from shared location during validation and
size calculation. This can result in incorrect empty space
calculation leading to out of bound access. Fix this by saving
the index values prior to comparison.

Change-Id: Idf501c31210ca97e93358b03fa24366aad3dd402
Signed-off-by: Kamal Agrawal <quic_kamaagra@quicinc.com>
Signed-off-by: Harshitha Sai Neelati <quic_hsaineel@quicinc.com>